Redone the way players are announced
Added a command for setting the announced bit to zero to test the plugin
Added a command that is used throught a socket to notify the server that
a player has registered.
Added a custom event that fires when a user registers on the site.
Moved the announcing from a thread to an event listener instead